rework: Новая организация каталогов, чтобы бот был дотупен из разных пакетов
функция, которая запускается локально теперь называется main.py

Теперь все хедлеры автоматически привязываются в боту в пакете bot
