+++ bot/__init__.py
@@ -1,0 +1,5 @@
+# Import this module like from bot import bot 
+# to get all handlers
+
+from .bot import bot
+from .handlers import *
